Route to the correct url regardless of service
When xi is appended to the host

https://dev.trade-tariff.service.gov.uk/xi/

URLs are generated like so:

https://dev.trade-tariff.service.gov.uk/api/v2/xi/v2/commodities/0702000007.json

This is very strange behaviour and probably something odd with the
faraday url generation method.

This change takes control of the url construction and does what we
expect:

https://dev.trade-tariff.service.gov.uk/xi/api/v2/commodities/0702000007.json

These routes are there successfully transformed via the frontend request
forwared and hit the xi backend.