Support format as a config option
We want to initialise resources with a properly deserialized response
format that will make it easier to load relationships in a simper
format.

This adds the deserialization functionality to the uktt gem which we
will consume in the duty calculator.

This will mean that we can do Commodity#import_measures more easily